1.027 11/12/2025

+++ main/config.h
@@ -47,7 +47,7 @@ typedef struct
 	float battery_scale;
 	float rtd_scale;
 	float duty_scale;
-	uint8_t duty_test;
+	int duty_test;
 } PCNT_CONFIG;
 
 typedef struct
